[wip] for tests for configure updates
