--- test/templater.js
@@ -1,98 +1,0 @@
-var assert = require( 'chai' ).assert;
-var fs = require( 'fs-extra' );
-
-var sugarcoat = require( '../index' );
-var fsp = require( '../lib/fs-promiser' );
-
-// suite( 'Templater: globPartials', function() { test( '', function() {});});
-// suite( 'Templater: readPartials', function() {});
-// suite( 'Templater: registerPartials', function() {});
-suite( 'Templater: File Prefixer', () => {
-
-    test( 'By default, assets are prefixed as .sugar-example. Output file is prefixed with "prefixed-"',  done => {
-
-        var config = {
-            settings: {
-                dest: './test/sugarcoat',
-                prefix: {
-                    assets: [
-                        './test/assert/prefixAssets.css'
-                    ]
-                }
-            },
-            sections: [
-                {
-                    title: 'CSS File',
-                    files: './test/assert/parseVarCode.css'
-                },
-                {
-                    title: 'CSS File 2',
-                    files: './test/assert/parseVarCode.css'
-
-                }
-            ]
-        };
-
-        var setupFiles = [
-            './test/assert/prefixAssets-assertDefault.css',
-            './test/sugarcoat/sugarcoat/css/prefixed-prefixAssets.css'
-        ];
-
-        sugarcoat( config )
-        .then( () => {
-
-            return Promise.all( setupFiles.map( fsp.readFile ))
-            .then( assets => {
-
-                assert.equal( assets[ 0 ], assets[ 1 ], 'prefixAssets-assertDefault.css matches');
-                done();
-            });
-        });
-    });
-
-    test( 'Prefixed output should use the selector designated in the config: `prefix.selector`', done => {
-
-        var config = {
-            settings: {
-                dest: './test/sugarcoat',
-                prefix: {
-                    assets: [
-                        './test/assert/prefixAssets.css'
-                    ],
-                    selector: '.designated-prefix'
-                }
-            },
-            sections: [
-                {
-                    title: 'CSS File',
-                    files: './test/assert/parseVarCode.css'
-                }
-            ]
-        };
-
-        var setupFiles = [
-            './test/assert/prefixAssets-assert.css',
-            './test/sugarcoat/sugarcoat/css/prefixed-prefixAssets.css'
-        ];
-
-        sugarcoat( config )
-        .then( () => {
-
-            return Promise.all( setupFiles.map( fsp.readFile ))
-            .then(  assets => {
-                assert.equal( assets[ 0 ], assets[ 1 ], 'prefixAssets-assert.css matches');
-                done();
-            });
-        });
-    });
-
-    teardown( done => {
-
-        fs.remove( './test/sugarcoat', err => {
-
-            if ( err ) return console.error( err );
-
-            done();
-        });
-    });
-});
